♠ :: rename optimizers -> methods
