+++ method/encode.py
