+++ method/hho_cvrp.py
@@ -2,11 +2,11 @@ import time
 import random
 import math
 
-import optimizers.mutate as mutate
-from optimizers.crossover import pmx
-from optimizers.encode import random_key
-from optimizers.local import two_opt_inverse
-from optimizers.local import two_opt_insertion
+import method.mutate as mutate
+from method.crossover import pmx
+from method.encode import random_key
+from method.local import two_opt_inverse
+from method.local import two_opt_insertion
 
 from model.solution import Solution
 from controller.benchmarks import concat_depot
